taro: Update kernel FCM version to 6 in device manifest
Specify kernel FCM version in the device manifest until
we fully switch to GKI i.e. With GKI, kernel release
string is set to a pre-defined value based on the version
of Android and is retrieved at run-time to get the right
kernel FCM version dynamically. Since we use a modified
version of kernel, it's not set to right value dynamically
and hence hard code it in device manifest.

Change-Id: Ica5e3eb8e9dc88951dd9a5ddaaf38f9109d5751e